+++ src/rapt.c
@@ -12,13 +12,13 @@
 #include "mvnorm.h"
 #include "vector_stats.h"
 
-mcmclib_rapt_suff* mcmclib_rapt_suff_alloc(int t0, int K, int dim) {
+mcmclib_rapt_suff* mcmclib_rapt_suff_alloc(size_t t0, size_t K, size_t dim) {
   mcmclib_rapt_suff* a = (mcmclib_rapt_suff*) malloc(sizeof(mcmclib_rapt_suff));
   a->t0 = t0;
 
   a->means = (gsl_vector**) malloc(K * sizeof(gsl_vector*));
   a->variances = (gsl_matrix**) malloc(K * sizeof(gsl_matrix*));
-  for(int k=0; k<K; k++) {
+  for(size_t k=0; k<K; k++) {
     a->means[k] = gsl_vector_alloc(dim);
     a->variances[k] = gsl_matrix_alloc(dim, dim);
   }
@@ -39,9 +39,9 @@ mcmclib_rapt_suff* mcmclib_rapt_suff_alloc(int t0, int K, int dim) {
 
 void mcmclib_rapt_suff_free(void* in_p) {
   mcmclib_rapt_suff* p = (mcmclib_rapt_suff*) in_p;
-  int K = p->n->size;
+  size_t K = p->n->size;
   gsl_vector_free(p->n);
-  for(int k=0; k< K; k++) {
+  for(size_t k=0; k< K; k++) {
     gsl_vector_free(p->means[k]);
     gsl_matrix_free(p->variances[k]);
   }
@@ -59,13 +59,13 @@ void mcmclib_rapt_update(mcmclib_amh* p);
 mcmclib_amh* mcmclib_rapt_alloc(gsl_rng* r,
 				distrfun_p logdistr, void* logdistr_data,
 				gsl_vector* x,
-				int t0,
+				size_t t0,
 				const gsl_matrix* sigma_whole,
-				int K,
+				size_t K,
 				gsl_matrix** sigma_local,
 				region_fun_t which_region,
 				void* which_region_data) {
-  int dim = x->size;
+  size_t dim = x->size;
 
   mcmclib_mh_q* q = mcmclib_rapt_q_alloc(r, logdistr, logdistr_data,
 					 sigma_whole, K, sigma_local,
@@ -100,10 +100,10 @@ void mcmclib_rapt_update_suff(mcmclib_amh* p) {
   mcmclib_mh* mh = p->mh;
   mcmclib_rapt_gamma* g = (mcmclib_rapt_gamma*) mh->q->gamma;
   gsl_vector* x = mh->x;
-  int k = g->which_region_x = g->which_region(g->which_region_data, x);
-  int fake_n = gsl_vector_get(s->n, k);
+  size_t k = g->which_region_x = g->which_region(g->which_region_data, x);
+  size_t fake_n = (size_t) gsl_vector_get(s->n, k);
   mcmclib_covariance_update(s->variances[k], s->means[k], &fake_n, x);
-  fake_n = p->n - 1;
+  fake_n = (size_t) (p->n - 1);
   mcmclib_covariance_update(s->global_variance, s->global_mean, &fake_n, x);
   gsl_vector_set(s->n, g->which_region_x, gsl_vector_get(s->n, g->which_region_x) + 1);
 }
